requirements.txt
groq>=0.4.0
streamlit>=1.28.0
pandas>=2.0.0
jsonschema>=4.19.0
python-dateutil>=2.8.0